requirements.txt
Pillow
flask